refactor: 💡 refactor tests to remove duplicated code
