+++ libs/libcommon/tests/conftest.py
@@ -1,10 +1,14 @@
 # SPDX-License-Identifier: Apache-2.0
 # Copyright 2022 The HuggingFace Authors.
 from pathlib import Path
+from typing import Iterator
 
 from environs import Env
 from pytest import fixture
 
+from libcommon.queue import _clean_queue_database
+from libcommon.resources import CacheMongoResource, QueueMongoResource
+from libcommon.simple_cache import _clean_cache_database
 from libcommon.storage import StrPath, init_cached_assets_dir
 
 # Import fixture modules as plugins
@@ -53,3 +57,27 @@ def metrics_mongo_host(env: Env) -> str:
 def cached_assets_directory(tmp_path: Path) -> StrPath:
     cached_assets_directory = tmp_path / "cached-assets"
     return init_cached_assets_dir(cached_assets_directory)
+
+
+@fixture
+def queue_mongo_resource(queue_mongo_host: str) -> Iterator[QueueMongoResource]:
+    database = "datasets_server_queue_test"
+    host = queue_mongo_host
+    if "test" not in database:
+        raise ValueError("Test must be launched on a test mongo database")
+    with QueueMongoResource(database=database, host=host, server_selection_timeout_ms=3_000) as queue_mongo_resource:
+        if not queue_mongo_resource.is_available():
+            raise RuntimeError("Mongo resource is not available")
+        yield queue_mongo_resource
+        _clean_queue_database()
+
+
+@fixture
+def cache_mongo_resource(cache_mongo_host: str) -> Iterator[CacheMongoResource]:
+    database = "datasets_server_cache_test"
+    host = cache_mongo_host
+    if "test" not in database:
+        raise ValueError("Test must be launched on a test mongo database")
+    with CacheMongoResource(database=database, host=host) as cache_mongo_resource:
+        yield cache_mongo_resource
+        _clean_cache_database()
